+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - PASSBOLT-2534: As LU I should not be able to copy to clipboard empty login/url
 - PASSBOLT-2017: As LU when I save a password (create/edit) the dialog shouldn't persist until the request is processed by the API
 - PASSBOLT-3063: Fix appjs base url and subfolder
+- PASSBOLT-3024: As LU I can't access the theme manager screen via url
 
 ### Added
 - PASSBOLT-2972: As LU I should be able to delete multiple passwords in bulk